Some compiler supports adding inline assembly code using #asm/#endasm
directives. For instance:

    #asm
        sll     a
        jrc     sub_1f      ; subtract $1f if A.x msb is set
        ret
    sub_1f:
        xor     a,#$1f
    #endasm

Another way is to use #pragma asm/endasm directives:

    #pragma asm
        XREF     _my_var
        PUSH     #LOW(my_func)
        PUSH     _my_var+0
        RETF
    App_CallEnd:
    #pragma endasm

This commit disables parsing of these section, much like when using the
keywords *INDENT-OFF* / *INDENT-ON*.

The way we detect the end of a section is limited. Since parsing is disabled, we
cannot rely on it. Instead we do a simple string search on either the word
'#endasm' or the words '#pragma' and 'endasm'.

Ideally the search should use regular expression, and the section begin/end
keywords should be user configurable.

----------------------
Pull Request ChangeLog
----------------------
- xxxxxxx    This commit

  * Remove unnecessary check for #endasm / #pragma endasm

- bfa3991    Original pull request

+++ src/args.h
@@ -19,9 +19,9 @@
 class Args
 {
 protected:
-   int   m_count;
-   char  **m_values;
-   UINT8 *m_used;      /* array of bits */
+   size_t m_count;
+   char   **m_values;
+   UINT8  *m_used;      /* array of bits */
 
 public:
    /**
@@ -63,28 +63,28 @@ public:
    const char *Param(const char *token);
 
    /**
-    * Similiar to arg_param, but can iterate over all matches.
+    * Similar to arg_param, but can iterate over all matches.
     * Set index to 0 before the first call.
     *
     * @param token   The token string to match
     * @param idx     Pointer to the index that you initialized to 0
     * @return        NULL or the pointer to the string.
     */
-   const char *Params(const char *token, int &index);
+   const char *Params(const char *token, size_t &index);
 
    /**
     * Marks an argument as being used.
     *
     * @param idx  The index of the argument
     */
-   void SetUsed(int idx);
+   void SetUsed(size_t idx);
 
    /**
     * Gets whether an argument has been used, by index.
     *
     * @param idx  The index of the argument
     */
-   bool GetUsed(int idx);
+   bool GetUsed(size_t idx);
 
    /**
     * This function retrieves all unused parameters.
@@ -94,7 +94,7 @@ public:
     * @param idx  Pointer to the index
     * @return     NULL (done) or the pointer to the string
     */
-   const char *Unused(int &idx);
+   const char *Unused(size_t &idx);
 
    /**
     * Takes text and splits it into arguments.
@@ -108,7 +108,7 @@ public:
     * @param num_args   The number of items in args
     * @return           The number of arguments parsed (always <= num_args)
     */
-   static int SplitLine(char *text, char *args[], int num_args);
+   static size_t SplitLine(char *text, char *args[], size_t num_args);
 };
 
 #endif /* ARGS_H_INCLUDED */
